backend/requirements.txt
python-dotenv
google-generativeai
opencv-python
numpy
pillow 